Cleanup: Moving and renaming fxns in symtext module
