+++ molsym/symtext/character_table.py
@@ -1,5 +1,156 @@
 import numpy as np
-from .symel import *
+from .point_group import PointGroup
+from ..symtools import *
+import re
+
+class CharacterTable():
+    def __init__(self, pg, irreps, classes, class_orders, chars, irrep_dims) -> None:
+        self.name = pg
+        self.irreps = irreps
+        self.classes = classes
+        self.class_orders = class_orders
+        self.characters = chars
+        self.irrep_dims = irrep_dims
+    def __repr__(self) -> str:
+        return f"Character Table for {self.name}\nIrreps: {self.irreps}\nClasses: {self.classes}\nCharacters:\n{self.characters}\n"
+    def __eq__(self, other):
+        if len(self.irreps) == len(other.irreps) and len(self.classes) == len(other.classes) and np.shape(self.characters)==np.shape(other.characters):
+            return (self.irreps == other.irreps).all() and (self.classes == other.classes).all() and np.isclose(self.characters,other.characters,atol=1e-10).all()
+        else:
+            return False
+
+def pg_to_chartab(PG):
+    pg = PointGroup.from_string(PG)
+    irreps = []
+    if pg.family == "C":
+        if pg.subfamily == "s":
+            irreps = ["A'","A''"]
+            classes = ["E", "sigma_h"]
+            chars = np.array([[1.0, 1.0], [1.0, -1.0]])
+        elif pg.subfamily == "i":
+            irreps = ["Ag","Au"]
+            classes = ["E", "i"]
+            chars = np.array([[1.0, 1.0], [1.0, -1.0]])
+        elif pg.subfamily == "v":
+            irreps, classes, chars = Cnv_irr(pg.n)
+        elif pg.subfamily == "h":
+            irreps, classes, chars = Cnh_irr(pg.n)
+        else:
+            #irreps, classes, chars = Cn_irrmat(pg.n)
+            irreps, classes, chars = Cn_irr_complex(pg.n)
+    elif pg.family == "D":
+        if pg.subfamily == "d":
+            irreps, classes, chars = Dnd_irr(pg.n)
+        elif pg.subfamily == "h":
+            irreps, classes, chars = Dnh_irr(pg.n)
+        else:
+            irreps, classes, chars = Dn_irr(pg.n)
+    elif pg.family == "S":
+        irreps, classes, chars = Sn_irr_complex(pg.n)
+    else:
+        cp3 = np.cos(np.pi/3)
+        pr5 = 0.5*(1.0+np.sqrt(5.0))
+        mr5 = 0.5*(1.0-np.sqrt(5.0))
+        if pg.family == "T":
+            if pg.subfamily == "h":
+                irreps, classes, chars = (["Ag","Au","Eg","Eu","Tg","Tu"],
+                 ["E", "4C_3", "4C_3^2", "3C_2", "i", "S_6", "S_6^5", "3sigma_h"],
+                 np.array(
+                 [[1.0,  1.0,  1.0,  1.0,  1.0,  1.0,  1.0,  1.0],
+                  [1.0,  1.0,  1.0,  1.0, -1.0, -1.0, -1.0, -1.0],
+                  [2.0,  cp3,  cp3,  2.0,  2.0,  cp3,  cp3,  1.0],
+                  [2.0,  cp3,  cp3,  2.0, -2.0, -cp3, -cp3, -1.0],
+                  [3.0,  0.0,  0.0, -1.0,  1.0,  0.0,  0.0, -1.0],
+                  [3.0,  0.0,  0.0, -1.0, -1.0,  0.0,  0.0,  1.0]]))
+            elif pg.subfamily == "d":
+                irreps, classes, chars = (["A1","A2","E","T1","T2"],
+                 ["E", "8C_3", "3C_2", "6S_4", "6sigma_d"],
+                 np.array(
+                 [[1.0,  1.0,  1.0,  1.0,  1.0],
+                  [1.0,  1.0,  1.0, -1.0, -1.0],
+                  [2.0, -1.0,  2.0,  0.0,  0.0],
+                  [3.0,  0.0, -1.0,  1.0, -1.0],
+                  [3.0,  0.0, -1.0, -1.0,  1.0]]))
+            else:
+                irreps, classes, chars = (["A","E","T"],
+                 ["E", "4C_3", "4C_3^2", "3C_2"],
+                 np.array(
+                 [[1.0,  1.0,  1.0,  1.0],
+                  [2.0,  cp3,  cp3,  2.0],
+                  [3.0,  0.0,  0.0, -1.0]]))
+        elif pg.family == "O":
+            if pg.subfamily == "h":
+                irreps, classes, chars = (["A1g","A2g","Eg","T1g","T2g","A1u","A2u","Eu","T1u","T2u"],
+                 ["E", "8C_3", "6C_2", "6C_4", "3C_2", "i", "6S_4", "8S_6", "3sigma_h", "6sigma_d"],
+                 np.array(
+                 [[1.0,  1.0,  1.0,  1.0,  1.0,  1.0,  1.0,  1.0,  1.0,  1.0],
+                  [1.0,  1.0, -1.0, -1.0,  1.0,  1.0, -1.0,  1.0,  1.0, -1.0],
+                  [2.0, -1.0,  0.0,  0.0,  2.0,  2.0,  0.0, -1.0,  2.0,  0.0],
+                  [3.0,  0.0, -1.0,  1.0, -1.0,  3.0,  1.0,  0.0, -1.0, -1.0],
+                  [3.0,  0.0,  1.0, -1.0, -1.0,  3.0, -1.0,  0.0, -1.0,  1.0],
+                  [1.0,  1.0,  1.0,  1.0,  1.0, -1.0, -1.0, -1.0, -1.0, -1.0],
+                  [1.0,  1.0, -1.0, -1.0,  1.0, -1.0,  1.0, -1.0, -1.0,  1.0],
+                  [2.0, -1.0,  0.0,  0.0,  2.0, -2.0,  0.0,  1.0, -2.0,  0.0],
+                  [3.0,  0.0, -1.0,  1.0, -1.0, -3.0, -1.0,  0.0,  1.0,  1.0],
+                  [3.0,  0.0,  1.0, -1.0, -1.0, -3.0,  1.0,  0.0,  1.0, -1.0]]))
+            else:
+                irreps, classes, chars = (["A1","A2","E","T1","T2"],
+                 ["E", "6C_4", "3C_2", "8C_3", "6C_2"],
+                 np.array(
+                 [[1.0,  1.0,  1.0,  1.0,  1.0],
+                  [1.0, -1.0,  1.0,  1.0, -1.0],
+                  [2.0,  0.0,  2.0, -1.0,  0.0],
+                  [3.0,  1.0, -1.0,  0.0, -1.0],
+                  [3.0, -1.0, -1.0,  0.0,  1.0]]))
+        elif pg.family == "I":
+            if pg.subfamily == "h":
+                irreps, classes, chars = (["Ag","T1g","T2g","Gg","Hg","Au","T1u","T2u","Gu","Hu"],
+                 ["E", "12C_5", "12C_5^2", "20C_3", "15C_2", "i", "12S_10", "12S_10^3", "20S_6", "15sigma"],
+                 np.array(
+                 [[1.0,  1.0,  1.0,  1.0,  1.0,  1.0,  1.0,  1.0,  1.0,  1.0],
+                  [3.0,  pr5,  mr5,  0.0, -1.0,  3.0,  mr5,  pr5,  0.0, -1.0],
+                  [3.0,  mr5,  pr5,  0.0, -1.0,  3.0,  pr5,  mr5,  0.0, -1.0],
+                  [4.0, -1.0, -1.0,  1.0,  0.0,  4.0, -1.0, -1.0,  1.0,  0.0],
+                  [5.0,  0.0,  0.0, -1.0,  1.0,  5.0,  0.0,  0.0, -1.0,  1.0],
+                  [1.0,  1.0,  1.0,  1.0,  1.0, -1.0, -1.0, -1.0, -1.0, -1.0],
+                  [3.0,  pr5,  mr5,  0.0, -1.0, -3.0, -mr5, -pr5,  0.0,  1.0],
+                  [3.0,  mr5,  pr5,  0.0, -1.0, -3.0, -pr5, -mr5,  0.0,  1.0],
+                  [4.0, -1.0, -1.0,  1.0,  0.0, -4.0,  1.0,  1.0, -1.0,  0.0],
+                  [5.0,  0.0,  0.0, -1.0,  1.0, -5.0,  0.0,  0.0,  1.0, -1.0]]))
+            else:
+                irreps, classes, chars = (["A","T1","T2","G","H"],
+                 ["E", "12C_5", "12C_5^2", "20C_3", "15C_2"],
+                 np.array(
+                 [[1.0,  1.0,  1.0,  1.0,  1.0],
+                  [3.0,  pr5,  mr5,  0.0, -1.0],
+                  [3.0,  mr5,  pr5,  0.0, -1.0],
+                  [4.0, -1.0, -1.0,  1.0,  0.0],
+                  [5.0,  0.0,  0.0, -1.0,  1.0]]))
+        else:
+            raise Exception(f"An invalid point group has been given or unexpected parsing of the point group string has occured: {pg.str}")
+    class_orders = grab_class_orders(classes)
+    irr_dims = {}
+    for (irr_idx,irrep) in enumerate(irreps):
+        if pg.n == 1:
+            irr_dims[irrep] = int(chars[0])
+        else:
+            irr_dims[irrep] = int(np.real(chars[irr_idx, 0]))
+    return CharacterTable(PG, irreps, classes, class_orders, chars, irr_dims)
+
+def grab_class_orders(classes):
+    ncls = len(classes)
+    class_orders = np.zeros(ncls)
+    for i in range(ncls): # = 1:ncls
+        class_orders[i] = grab_order(classes[i])
+    return class_orders
+
+def grab_order(class_str):
+    regex = r"^(\d+)"
+    m = re.match(regex, class_str)
+    if m is not None:
+        return int(m.groups()[0])
+    else:
+        return 1
 
 def c_class_string(classes, pre, r, s):
     "Pushes class string to 'classes' for rotations, but ignores superscript if s is one"
